% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacamp_export.R
\name{courses}
\alias{courses}
\alias{users}
\alias{missing_courses}
\title{Extract tidy data from DataCamp exports}
\usage{
courses(x)

users(x)

missing_courses(x)
}
\arguments{
\item{x}{An object of class \code{"datacamp_export"}.}
}
\description{
Extract tidy data from DataCamp exports
}
