% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datacamp.R
\name{read_datacamp}
\alias{read_datacamp}
\title{Read the contents of a DataCamp group export into R.}
\usage{
read_datacamp(path)
}
\arguments{
\item{path}{Either the zip file downloaded from the website (must end in
\code{.zip}), or the unzipped directory containing the exported files.}
}
\description{
Read the contents of a DataCamp group export into R.
}
