% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_group_export.R
\name{read_group_export}
\alias{read_group_export}
\title{Read the contents of a DataCamp group export into R}
\usage{
read_group_export(path)
}
\arguments{
\item{path}{Either the zip file downloaded from the website (must end in
\code{.zip}), or the unzipped directory containing the exported files.}
}
\description{
Read the contents of a DataCamp group export into R
}
